==> Nebraska/WaterUse_AggregatedArea/0_NEAggregatedDataPreprocess.ipynb <==
# NE Aggregated Data
Preprocessing input data for a smoother upload experience of the state data to the WaDE 2.0 database.
Using geopandas to read in shp file, and coverting to WKT for ReportingUnit geometry.

Notes:
- Date Updated: 05/02/2022
- Source data is missing some basic inputs, will have to hard code those there and match up with the timeseries.
- After seven different timeseries data.
- Will use basin data, and subbasin data.

#Needed Libararies
import os
import numpy as np
import pandas as pd
from datetime import datetime
import geopandas as gpd # the library that lets us read in shapefiles
pd.set_option('display.max_columns', 999)  # How to display all columns of a Pandas DataFrame in Jupyter Notebook

#Setting work directory, reading inputs, creating dataframe
workingDir = "G:/Shared drives/WaDE Data/Nebraska/AggregatedAmounts"
os.chdir(workingDir)

## Basin - Water Use Data

#CSV input file
fileInput = "RawInputData/INSIGHT_FinalBasinData_20151102_input.csv"
df_ba = pd.read_csv(fileInput)
print(len(df_ba))
df_ba.head(1)

# 1, Depletion_Annual_Total_Surface Water
df1_ba_DATS = pd.DataFrame(index=df_ba.index)

# Variable Info
df1_ba_DATS['in_VariableSpecificCV'] = "Depletion_Annual_Total_Surface Water"

# WaterSource Info
df1_ba_DATS['in_WaterSourceTypeCV'] = "Surface Water"

# ReportingUnit Info
df1_ba_DATS['in_ReportingUnitName'] = df_ba['Basin']
df1_ba_DATS['in_ReportingUnitNativeID'] = df_ba['BasinID']
df1_ba_DATS['in_ReportingUnitTypeCV'] = "Basin"

# AggregatedAmounts Info
df1_ba_DATS['in_ReportYearCV'] = df_ba['Year']
df1_ba_DATS['in_Amount'] = df_ba['SWDTotal_Annual']
df1_ba_DATS['in_BeneficialUseCategory'] = "Total"

print(len(df1_ba_DATS))
df1_ba_DATS.head(1)

# 2, Basin, Depletion_Annual_Total_Groundwater
df2_ba_DATG = pd.DataFrame(index=df_ba.index)

# Variable Info
df2_ba_DATG['in_VariableSpecificCV'] = "Depletion_Annual_Total_Groundwater"

# WaterSource Info
df2_ba_DATG['in_WaterSourceTypeCV'] = "Groundwater"

# ReportingUnit Info
df2_ba_DATG['in_ReportingUnitName'] = df_ba['Basin']
df2_ba_DATG['in_ReportingUnitNativeID'] = df_ba['BasinID']
df2_ba_DATG['in_ReportingUnitTypeCV'] = "Basin"

# AggregatedAmounts Info
df2_ba_DATG['in_ReportYearCV'] = df_ba['Year']
df2_ba_DATG['in_Amount'] = df_ba['GWDP_Annual']
df2_ba_DATG['in_BeneficialUseCategory'] = "Total"

print(len(df2_ba_DATG))
df2_ba_DATG.head(1)

# 3, Basin, Consumptive Use_Annual_Total_Groundwater
d3f_ba_CATG = pd.DataFrame(index=df_ba.index)

# Variable Info
d3f_ba_CATG['in_VariableSpecificCV'] = "Consumptive Use_Annual_Total_Groundwater"

# WaterSource Info
d3f_ba_CATG['in_WaterSourceTypeCV'] = "Groundwater"

# ReportingUnit Info
d3f_ba_CATG['in_ReportingUnitName'] = df_ba['Basin']
d3f_ba_CATG['in_ReportingUnitNativeID'] = df_ba['BasinID']
d3f_ba_CATG['in_ReportingUnitTypeCV'] = "Basin"

# AggregatedAmounts Info
d3f_ba_CATG['in_ReportYearCV'] = df_ba['Year']
d3f_ba_CATG['in_Amount'] = df_ba['GWCTotal_Annual']
d3f_ba_CATG['in_BeneficialUseCategory'] = "Total"

print(len(d3f_ba_CATG))
d3f_ba_CATG.head(1)

# 4, Basin, Consumptive Use_Annual_Irrigation_Groundwater
df4_ba_CAIrG = pd.DataFrame(index=df_ba.index)

# Variable Info
df4_ba_CAIrG['in_VariableSpecificCV'] = "Consumptive Use_Annual_Irrigation_Groundwater"

# WaterSource Info
df4_ba_CAIrG['in_WaterSourceTypeCV'] = "Groundwater"

# ReportingUnit Info
df4_ba_CAIrG['in_ReportingUnitName'] = df_ba['Basin']
df4_ba_CAIrG['in_ReportingUnitNativeID'] = df_ba['BasinID']
df4_ba_CAIrG['in_ReportingUnitTypeCV'] = "Basin"

# AggregatedAmounts Info
df4_ba_CAIrG['in_ReportYearCV'] = df_ba['Year']
df4_ba_CAIrG['in_Amount'] = df_ba['GWCIrrigation_Annual']
df4_ba_CAIrG['in_BeneficialUseCategory'] = "Irrigation"

print(len(df4_ba_CAIrG))
df4_ba_CAIrG.head(1)

# 5, Basin, Consumptive Use_Annual_Municipal_Groundwater
df5_ba_CAMG = pd.DataFrame(index=df_ba.index)

# Variable Info
df5_ba_CAMG['in_VariableSpecificCV'] = "Consumptive Use_Annual_Municipal_Groundwater"

# WaterSource Info
df5_ba_CAMG['in_WaterSourceTypeCV'] = "Groundwater"

# ReportingUnit Info
df5_ba_CAMG['in_ReportingUnitName'] = df_ba['Basin']
df5_ba_CAMG['in_ReportingUnitNativeID'] = df_ba['BasinID']
df5_ba_CAMG['in_ReportingUnitTypeCV'] = "Basin"

# AggregatedAmounts Info
df5_ba_CAMG['in_ReportYearCV'] = df_ba['Year']
df5_ba_CAMG['in_Amount'] = df_ba['GWCMunicipal_Annual']
df5_ba_CAMG['in_BeneficialUseCategory'] = "Municipal"

print(len(df5_ba_CAMG))
df5_ba_CAMG.head(1)

# 6, Basin, Consumptive Use_Annual_Industrial_Groundwater
df6_ba_CAInG = pd.DataFrame(index=df_ba.index)

# Variable Info
df6_ba_CAInG['in_VariableSpecificCV'] = "Consumptive Use_Annual_Industrial_Groundwater"

# WaterSource Info
df6_ba_CAInG['in_WaterSourceTypeCV'] = "Groundwater"

# ReportingUnit Info
df6_ba_CAInG['in_ReportingUnitName'] = df_ba['Basin']
df6_ba_CAInG['in_ReportingUnitNativeID'] = df_ba['BasinID']
df6_ba_CAInG['in_ReportingUnitTypeCV'] = "Basin"

# AggregatedAmounts Info
df6_ba_CAInG['in_ReportYearCV'] = df_ba['Year']
df6_ba_CAInG['in_Amount'] = df_ba['GWCIndustrial_Annual']
df6_ba_CAInG['in_BeneficialUseCategory'] = "Industrial"

print(len(df6_ba_CAInG))
df6_ba_CAInG.head(1)

# 7, Basin, Demand_Annual_Total_Surface Water
df7_ba_DeATS = pd.DataFrame(index=df_ba.index)

# Variable Info
df7_ba_DeATS['in_VariableSpecificCV'] = "Demand_Annual_Total_Surface Water"

# WaterSource Info
df7_ba_DeATS['in_WaterSourceTypeCV'] = "Surface Water"

# ReportingUnit Info
df7_ba_DeATS['in_ReportingUnitName'] = df_ba['Basin']
df7_ba_DeATS['in_ReportingUnitNativeID'] = df_ba['BasinID']
df7_ba_DeATS['in_ReportingUnitTypeCV'] = "Basin"

# AggregatedAmounts Info
df7_ba_DeATS['in_ReportYearCV'] = df_ba['Year']
df7_ba_DeATS['in_Amount'] = df_ba['SWDemandTotal_Annual']
df7_ba_DeATS['in_BeneficialUseCategory'] = "Total"

print(len(df7_ba_DeATS))
df7_ba_DeATS.head(1)

# Concatenate Basin DataFrames
frames = [df1_ba_DATS, df2_ba_DATG, d3f_ba_CATG, df4_ba_CAIrG, df5_ba_CAMG, df6_ba_CAInG, df7_ba_DeATS]
df_BasinOUT = pd.concat(frames)

print(len(df_BasinOUT))
df_BasinOUT.head(1)

# variable check
df_BasinOUT['in_VariableSpecificCV'].unique()

## Subbasin - Water Use Data

#CSV input file
fileInput = "RawInputData/INSIGHT_FinalSubBasinData_20151102_input.csv"
df_sub = pd.read_csv(fileInput)
print(len(df_sub))
df_sub.head(1)

# 1, Depletion_Annual_Total_Surface Water
df1_sub_DATS = pd.DataFrame(index=df_sub.index)

# Variable Info
df1_sub_DATS['in_VariableSpecificCV'] = "Depletion_Annual_Total_Surface Water"

# WaterSource Info
df1_sub_DATS['in_WaterSourceTypeCV'] = "Surface Water"

# ReportingUnit Info
df1_sub_DATS['in_ReportingUnitName'] = df_sub['Subbasin']
df1_sub_DATS['in_ReportingUnitNativeID'] = df_sub['SubID']
df1_sub_DATS['in_ReportingUnitTypeCV'] = "Subbasin"

# AggregatedAmounts Info
df1_sub_DATS['in_ReportYearCV'] = df_sub['Year']
df1_sub_DATS['in_Amount'] = df_sub['SWDTotal_Annual']
df1_sub_DATS['in_BeneficialUseCategory'] = "Total"

print(len(df1_sub_DATS))
df1_sub_DATS.head(1)

# 2, Basin, Depletion_Annual_Total_Groundwater
df2_sub_DATG = pd.DataFrame(index=df_sub.index)

# Variable Info
df2_sub_DATG['in_VariableSpecificCV'] = "Depletion_Annual_Total_Groundwater"

# WaterSource Info
df2_sub_DATG['in_WaterSourceTypeCV'] = "Groundwater"

# ReportingUnit Info
df2_sub_DATG['in_ReportingUnitName'] = df_sub['Subbasin']
df2_sub_DATG['in_ReportingUnitNativeID'] = df_sub['SubID']
df2_sub_DATG['in_ReportingUnitTypeCV'] = "Subbasin"

# AggregatedAmounts Info
df2_sub_DATG['in_ReportYearCV'] = df_sub['Year']
df2_sub_DATG['in_Amount'] = df_sub['GWDP_Annual']
df2_sub_DATG['in_BeneficialUseCategory'] = "Total"

print(len(df2_sub_DATG))
df2_sub_DATG.head(1)

# 3, Basin, Consumptive Use_Annual_Total_Groundwater
d3f_sub_CATG = pd.DataFrame(index=df_sub.index)

# Variable Info
d3f_sub_CATG['in_VariableSpecificCV'] = "Consumptive Use_Annual_Total_Groundwater"

# WaterSource Info
d3f_sub_CATG['in_WaterSourceTypeCV'] = "Groundwater"

# ReportingUnit Info
d3f_sub_CATG['in_ReportingUnitName'] = df_sub['Subbasin']
d3f_sub_CATG['in_ReportingUnitNativeID'] = df_sub['SubID']
d3f_sub_CATG['in_ReportingUnitTypeCV'] = "Subbasin"

# AggregatedAmounts Info
d3f_sub_CATG['in_ReportYearCV'] = df_sub['Year']
d3f_sub_CATG['in_Amount'] = df_sub['GWCTotal_Annual']
d3f_sub_CATG['in_BeneficialUseCategory'] = "Total"

print(len(d3f_sub_CATG))
d3f_sub_CATG.head(1)

# 4, Basin, Consumptive Use_Annual_Irrigation_Groundwater
df4_sub_CAIrG = pd.DataFrame(index=df_sub.index)

# Variable Info
df4_sub_CAIrG['in_VariableSpecificCV'] = "Consumptive Use_Annual_Irrigation_Groundwater"

# WaterSource Info
df4_sub_CAIrG['in_WaterSourceTypeCV'] = "Groundwater"

# ReportingUnit Info
df4_sub_CAIrG['in_ReportingUnitName'] = df_sub['Subbasin']
df4_sub_CAIrG['in_ReportingUnitNativeID'] = df_sub['SubID']
df4_sub_CAIrG['in_ReportingUnitTypeCV'] = "Subbasin"

# AggregatedAmounts Info
df4_sub_CAIrG['in_ReportYearCV'] = df_sub['Year']
df4_sub_CAIrG['in_Amount'] = df_sub['GWCIrrigation_Annual']
df4_sub_CAIrG['in_BeneficialUseCategory'] = "Irrigation"

print(len(df4_sub_CAIrG))
df4_sub_CAIrG.head(1)

# 5, Basin, Consumptive Use_Annual_Municipal_Groundwater
df5_sub_CAMG = pd.DataFrame(index=df_sub.index)

# Variable Info
df5_sub_CAMG['in_VariableSpecificCV'] = "Consumptive Use_Annual_Municipal_Groundwater"

# WaterSource Info
df5_sub_CAMG['in_WaterSourceTypeCV'] = "Groundwater"

# ReportingUnit Info
df5_sub_CAMG['in_ReportingUnitName'] = df_sub['Subbasin']
df5_sub_CAMG['in_ReportingUnitNativeID'] = df_sub['SubID']
df5_sub_CAMG['in_ReportingUnitTypeCV'] = "Subbasin"

# AggregatedAmounts Info
df5_sub_CAMG['in_ReportYearCV'] = df_sub['Year']
df5_sub_CAMG['in_Amount'] = df_sub['GWCMunicipal_Annual']
df5_sub_CAMG['in_BeneficialUseCategory'] = "Municipal"

print(len(df5_sub_CAMG))
df5_sub_CAMG.head(1)

# 6, Basin, Consumptive Use_Annual_Industrial_Groundwater
df6_sub_CAInG = pd.DataFrame(index=df_sub.index)

# Variable Info
df6_sub_CAInG['in_VariableSpecificCV'] = "Consumptive Use_Annual_Industrial_Groundwater"

# WaterSource Info
df6_sub_CAInG['in_WaterSourceTypeCV'] = "Groundwater"

# ReportingUnit Info
df6_sub_CAInG['in_ReportingUnitName'] = df_sub['Subbasin']
df6_sub_CAInG['in_ReportingUnitNativeID'] = df_sub['SubID']
df6_sub_CAInG['in_ReportingUnitTypeCV'] = "Subbasin"

# AggregatedAmounts Info
df6_sub_CAInG['in_ReportYearCV'] = df_sub['Year']
df6_sub_CAInG['in_Amount'] = df_sub['GWCIndustrial_Annual']
df6_sub_CAInG['in_BeneficialUseCategory'] = "Industrial"

print(len(df6_sub_CAInG))
df6_sub_CAInG.head(1)

# 7, Basin, Demand_Annual_Total_Surface Water
df7_sub_DeATS = pd.DataFrame(index=df_sub.index)

# Variable Info
df7_sub_DeATS['in_VariableSpecificCV'] = "Demand_Annual_Total_Surface Water"

# WaterSource Info
df7_sub_DeATS['in_WaterSourceTypeCV'] = "Surface Water"

# ReportingUnit Info
df7_sub_DeATS['in_ReportingUnitName'] = df_sub['Subbasin']
df7_sub_DeATS['in_ReportingUnitNativeID'] = df_sub['SubID']
df7_sub_DeATS['in_ReportingUnitTypeCV'] = "Subbasin"

# AggregatedAmounts Info
df7_sub_DeATS['in_ReportYearCV'] = df_sub['Year']
df7_sub_DeATS['in_Amount'] = df_sub['SWDemandTotal_Annual']
df7_sub_DeATS['in_BeneficialUseCategory'] = "Total"

print(len(df7_sub_DeATS))
df7_sub_DeATS.head(1)

# Concatenate Subbasin DataFrames
frames = [df1_sub_DATS, df2_sub_DATG, d3f_sub_CATG, df4_sub_CAIrG, df5_sub_CAMG, df6_sub_CAInG, df7_sub_DeATS]
df_SubbasinOUT = pd.concat(frames)

print(len(df_SubbasinOUT))
df_SubbasinOUT.head(1)

# variable check
df_SubbasinOUT['in_ReportingUnitTypeCV'].unique()

## Concatenate Basin & Subbasin DataFrames

# Concatenate Subbasin DataFrames
frames = [df_BasinOUT, df_SubbasinOUT]
df_out = pd.concat(frames).reset_index(drop=True)

print(len(df_out))
df_out.head()

## WaDE Custom Elements (due to missing info)

# Creating WaDE Custom water source native ID for easy water source identification
# ----------------------------------------------------------------------------------------------------

# Create temp WaterSourceNativeID dataframe of unique water source.
def assignWaterSourceNativeID(colrowValue):
    string1 = str(colrowValue)
    outstring = "WaDENE_WS" + string1
    return outstring

dfWaterSourceNativeID = pd.DataFrame()
dfWaterSourceNativeID['in_WaterSourceTypeCV'] = df_out['in_WaterSourceTypeCV']
dfWaterSourceNativeID = dfWaterSourceNativeID.drop_duplicates()

dftemp = pd.DataFrame(index=dfWaterSourceNativeID.index)
dftemp["Count"] = range(1, len(dftemp.index) + 1)
dfWaterSourceNativeID['in_WaterSourceNativeID'] = dftemp.apply(lambda row: assignWaterSourceNativeID(row['Count']), axis=1)

# ----------------------------------------------------------------------------------------------------

# Retreive WaDE Custom water source native ID
def retrieveWaterSourceNativeID(A):
    if (A == '') or (pd.isnull(A)):
        outList = ''
    else:
        ml = dfWaterSourceNativeID.loc[(dfWaterSourceNativeID['in_WaterSourceTypeCV'] == A), 'in_WaterSourceNativeID']
        if not (ml.empty):  # check if the series is empty
            outList = ml.iloc[0]
        else:
            outList = ''
    return outList

df_out['in_WaterSourceNativeID'] = df_out.apply(lambda row: retrieveWaterSourceNativeID(row['in_WaterSourceTypeCV']), axis=1)
df_out

## Shapefile Data

# Basin Shapefile Data
BasinSHPInput = "RawInputData/InsightShapeFiles/Basin.shp"
BasinShape = gpd.read_file(BasinSHPInput, crs="EPSG:4326")

print(len(BasinShape))
BasinShape.head(3)

# Basin output shapefile dataframe
columnList = ['in_ReportingUnitName', 
              'in_ReportingUnitNativeID',
              'in_ReportingUnitTypeCV',
              'in_Geomerty']
df_basin_shp = pd.DataFrame(columns=columnList, index=BasinShape.index)

df_basin_shp['in_ReportingUnitName'] = BasinShape['Basin']
df_basin_shp['in_ReportingUnitNativeID'] = BasinShape['BID']
df_basin_shp['in_ReportingUnitTypeCV'] = "Basin"
df_basin_shp['in_Geomerty'] = BasinShape['geometry']

print(len(df_basin_shp))
df_basin_shp.head(3)

# Subbasin Shapefile Data
SubbasinSHPInput = "RawInputData/InsightShapeFiles/Subbasin.shp"
SubbasinShape = gpd.read_file(SubbasinSHPInput)

print(len(SubbasinShape))
SubbasinShape.head(3)

# Subbasin output shapefile dataframe
columnList = ['in_ReportingUnitName', 
              'in_ReportingUnitNativeID',
              'in_ReportingUnitTypeCV',
              'in_Geomerty']
df_subbasin_shp = pd.DataFrame(columns=columnList, index=SubbasinShape.index)

df_subbasin_shp['in_ReportingUnitName'] = SubbasinShape['Subbasin']
df_subbasin_shp['in_ReportingUnitNativeID'] = SubbasinShape['SubID']
df_subbasin_shp['in_ReportingUnitTypeCV'] = "Subbasin"
df_subbasin_shp['in_Geomerty'] = SubbasinShape['geometry']

print(len(df_subbasin_shp))
df_subbasin_shp.head(3)

# Concatenate Basin & Subbasin DataFrames
frames = [df_basin_shp, df_subbasin_shp]
df_shape_out = pd.concat(frames)

print(len(df_shape_out))
df_shape_out.head()

### Inspect Output Data & Export

with pd.option_context('display.max_rows', None, 'display.max_columns', None):
    print(df_out.dtypes)

with pd.option_context('display.max_rows', None, 'display.max_columns', None):
    print(df_shape_out.dtypes)

# Export out to CSV.
df_out.to_csv('RawInputData/P_neAgMaster.csv', index=False) # The output.
df_shape_out.to_csv('RawInputData/P_neAgGeometry.csv', index=False) # The output geometry.
